requirements.txt
torch
transformers
datasets
accelerate
peft
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
matplotlib
tqdm
evaluate
python-dotenv
tensorboard
tiktoken
sentencepiece
rouge_score
nltk
streamlit